Refactor current version to v1.7
