+++ src/index.ts
@@ -1,5 +1,9 @@
-export { EosAction, EosAuthorization, EosPayload } from './interfaces'
-export * from './handlers/massiveeos'
-export * from './readers/mongo'
-export * from './readers/nodeos'
-export * from './readers/state-history'
+// Set default version here
+export * from './v1.7'
+
+// Export all versions for types support
+import * as v1_7 from './v1.7'
+
+export {
+  v1_7,
+}
