+++ src/v1.7/readers/nodeos/NodeosBlock.ts
