+++ src/v1.7/readers/state-history/interfaces.ts
@@ -1,0 +1,7 @@
+import { ActionReaderOptions } from 'demux'
+
+export interface StateHistoryPostgresActionReaderOptions extends ActionReaderOptions {
+  massiveConfig: any
+  dbSchema?: string
+  enablePgMonitor?: boolean
+}
